+++ fiberadmin.php
@@ -3,7 +3,7 @@
  * Plugin Name:       Fiber Admin
  * Plugin URI:        https://wordpress.org/plugins/fiber-admin/
  * Description:       💈 Another helpful admin with some extra functions for WordPress backend
- * Version:           1.5.3
+ * Version:           1.5.4
  * Requires at least: 5.2
  * Requires PHP:      7.2
  * Author:            Dao
@@ -25,9 +25,10 @@ if(!defined('ABSPATH')){
  * Definitions
  */
 
-define('FIBERADMIN_VERSION', '1.5.3');
+define('FIBERADMIN_VERSION', '1.5.4');
 define("FIBERADMIN_DIR", plugin_dir_path(__FILE__));
 define("FIBERADMIN_ASSETS_URL", plugin_dir_url(__FILE__) . 'assets/');
+define("FIBERADMIN_DEV_MODE", false);
 
 /**
  * Init Functions